backend/requirements.txt
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-WTF==1.0.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
phonenumbers==8.12.52
psycopg==3.0.15
psycopg-binary==3.0.15
pycparser==2.21
PyJWT==2.4.0
six==1.16.0
tzdata==2022.1
Werkzeug==2.2.0
WTForms==3.0.1
zipp==3.8.1
click==8.0.3
Flask==2.0.2
Flask-JWT-Extended==4.3.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
PyJWT==2.3.0
python-dotenv==0.19.2
Werkzeug==2.0.2
